+++ db/migrations/0014_overjoyed_blob.sql
@@ -1,0 +1,11 @@
+CREATE TABLE "service_image" (
+	"id" text PRIMARY KEY NOT NULL,
+	"service_id" text NOT NULL,
+	"url" text NOT NULL,
+	"pathname" text NOT NULL,
+	"created_at" timestamp DEFAULT now() NOT NULL,
+	"updated_at" timestamp DEFAULT now() NOT NULL
+);
+--> statement-breakpoint
+ALTER TABLE "service_image" ADD CONSTRAINT "service_image_service_id_service_id_fk" FOREIGN KEY ("service_id") REFERENCES "public"."service"("id") ON DELETE cascade ON UPDATE no action;--> statement-breakpoint
+CREATE INDEX "service_image_service_idx" ON "service_image" USING btree ("service_id");